introduce new accountidfor + reorganize registry traits
