+++ traits/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "hydradx-traits"
-version = "0.5.1"
+version = "0.6.0"
 description = "Shared traits"
 authors = ["GalacticCouncil"]
 edition = "2021"
