+++ traits/src/lib.rs
@@ -19,6 +19,8 @@
 #![allow(clippy::upper_case_acronyms)]
 
 pub mod pools;
+pub mod registry;
+pub use registry::*;
 
 use codec::{Decode, Encode};
 use frame_support::dispatch;
@@ -130,48 +132,6 @@ pub trait Resolver<AccountId, Intention, E> {
     /// Resolve intentions by either directly trading with each other or via AMM pool.
     /// Intention ```intention``` must be validated prior to call this function.
     fn resolve_matched_intentions(pair_account: &AccountId, intention: &Intention, matched: &[&Intention]);
-}
-
-pub trait Registry<AssetId, AssetName, Balance, Error> {
-    fn exists(name: AssetId) -> bool;
-
-    fn retrieve_asset(name: &AssetName) -> Result<AssetId, Error>;
-
-    fn create_asset(name: &AssetName, existential_deposit: Balance) -> Result<AssetId, Error>;
-
-    fn get_or_create_asset(name: AssetName, existential_deposit: Balance) -> Result<AssetId, Error> {
-        if let Ok(asset_id) = Self::retrieve_asset(&name) {
-            Ok(asset_id)
-        } else {
-            Self::create_asset(&name, existential_deposit)
-        }
-    }
-}
-
-pub trait ShareTokenRegistry<AssetId, AssetName, Balance, Error>: Registry<AssetId, AssetName, Balance, Error> {
-    fn retrieve_shared_asset(name: &AssetName, assets: &[AssetId]) -> Result<AssetId, Error>;
-
-    fn create_shared_asset(
-        name: &AssetName,
-        assets: &[AssetId],
-        existential_deposit: Balance,
-    ) -> Result<AssetId, Error>;
-
-    fn get_or_create_shared_asset(
-        name: AssetName,
-        assets: Vec<AssetId>,
-        existential_deposit: Balance,
-    ) -> Result<AssetId, Error> {
-        if let Ok(asset_id) = Self::retrieve_shared_asset(&name, &assets) {
-            Ok(asset_id)
-        } else {
-            Self::create_shared_asset(&name, &assets, existential_deposit)
-        }
-    }
-}
-
-pub trait AssetPairAccountIdFor<AssetId, AccountId> {
-    fn from_assets(asset_a: AssetId, asset_b: AssetId, identifier: &str) -> AccountId;
 }
 
 /// Handler used by AMM pools to perform some tasks when a new pool is created.
